feat(settings): add chat section
